pyproject.toml
[project]
name = "actiondocs"
version = "0.0.1"
description = "A GitHub Actions Markdown docs generator"
authors = [{name = "Pierre Nicolas Durette", email = "pndurette@gmail.com"}]
requires-python = ">=3.7"

dependencies = [
    "PyYAML==6.0",
    "Markdown==3.4.1"
]

[project.optional-dependencies]
tests = [
    "pytest ~= 7.2",
]